Move toElementsWithMappedProps to utils folder
